Migrated calls to vcastqt into vcastAdapter
